更新版本号到v1.7,添加新功能,优化代码

+++ wmzz_ban_cron.php
@@ -43,34 +43,51 @@ function cron_wmzz_ban()
     $now = strtotime(date('Y-m-d'));
     $y = $m->query("SELECT * FROM `" . DB_PREFIX . "wmzz_ban` WHERE `nextdo` <= '{$now}' LIMIT {$s['limit']}");
     while ($x = $m->fetch_array($y)) {
-        $r = wmzz_ban_getTime($now, $x['date'], $x['day']);
-        if ($r >= '0') {
-            $bduss = misc::getCookie($x['pid']);
-            $c = new wcurl('http://tieba.baidu.com/pmc/blockid');
-            $c->addcookie('BDUSS=' . $bduss);
-            $option = array(
-                'day' => $r,
-                'fid' => misc::getFid($x['tieba']),
-                'tbs' => misc::getTbs($x['uid'], $bduss),
-//                'ie' => 'utf-8',
-                'ie' => 'gbk',
-                'user_name[]' => '',
-                'nick_name[]' => '',
-                'portrait[]' => $x['portrait'],
-                'pid[]' => '',
-                'reason' => $x['msg']
-            );
-            $res = $c->post($option);
-            $res = json_decode($res, TRUE);
-            if ($res['errno'] == 0) {
-                $next = $now + ($r * 86400);
-                $m->query("UPDATE `" . DB_PREFIX . "wmzz_ban` SET `nextdo` = '{$next}' WHERE `id` = '{$x['id']}'");
+        $is_open = option::uget('wmzz_ban_enable', $x['uid']);
+        if (!$is_open) {
+            option::uset('wmzz_ban_enable', 'on', $x['uid']);
+            $is_open = true;
+        } else {
+            if ($is_open == "on") {
+                $is_open = true;
+            } else if ($is_open == "off") {
+                $is_open = false;
             }
+        }
+        $r = wmzz_ban_getTime($now, $x['date'], $x['day']);
+        if ($is_open) {
+            if ($r >= '0') {
+                $bduss = misc::getCookie($x['pid']);
+                $c = new wcurl('http://tieba.baidu.com/pmc/blockid');
+                $c->addcookie('BDUSS=' . $bduss);
+                $option = array(
+                    'day' => $r,
+                    'fid' => misc::getFid($x['tieba']),
+                    'tbs' => misc::getTbs($x['uid'], $bduss),
+//                'ie' => 'utf-8',
+                    'ie' => 'gbk',
+                    'user_name[]' => '',
+                    'nick_name[]' => '',
+                    'portrait[]' => $x['portrait'],
+                    'pid[]' => '',
+                    'reason' => $x['msg']
+                );
+                $res = $c->post($option);
+                $res = json_decode($res, TRUE);
+                if ($res['errno'] == 0) {
+                    $next = $now + ($r * 86400);
+                    $m->query("UPDATE `" . DB_PREFIX . "wmzz_ban` SET `nextdo` = '{$next}' WHERE `id` = '{$x['id']}'");
+                }
 //            else if ($res['errno'] == 74) {    //用户名不存在   224011 需要验证码
 //                $m->query("DELETE FROM `" . DB_PREFIX . "wmzz_ban` WHERE `id` = '{$x['id']}'");
 //            }
+            } else {
+                $m->query("DELETE FROM `" . DB_PREFIX . "wmzz_ban` WHERE `id` = '{$x['id']}'");
+            }
         } else {
-            $m->query("DELETE FROM `" . DB_PREFIX . "wmzz_ban` WHERE `id` = '{$x['id']}'");
+            $next = $now + ($r * 86400);
+            $m->query("UPDATE `" . DB_PREFIX . "wmzz_ban` SET `nextdo` = '{$next}' WHERE `id` = '{$x['id']}'");
+
         }
     }
 }
